start of allClasses
some methods and initializes need to be further done, I'm not exactly
sure how to do the two-dimensional array in ruby and the players need a
good data structure to store their regiments that also allows for
deletion of the regiments (or alternitively the game just goes over all
them and ignores the ones thart are dead. Definitely would like some
feedback as I continue to work on this, and learn from whatever either
of you know about ruby in this regard.